+++ .gitignore
@@ -1,2 +1,2 @@
 node_modules
-bundle.js
+client/dist/bundle.js
