+++ client/src/_tests_/App.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import App from '../components/App.jsx';
+
+describe('Rendering first component', () => {
+  it('renders without crashing', () => {
+   shallow(<App/>)
+  })
+})
