+++ client/src/index.jsx
@@ -1,17 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-    }
-  }
-  render () {
-    return (
-      <div>
-        <h1>Hello! From React</h1>
-      </div>)
-  }
-}
+import App from './components/App.jsx';
+
 ReactDOM.render(<App />, document.getElementById('app'));
