+++ package.json
@@ -11,7 +11,13 @@
     "seed": "node ./server/models/seed.js",
     "start": "nodemon server/index.js",
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/index.js"
+    "server-dev": "nodemon server/index.js",
+    "test": "jest"
+  },
+  "jest": {
+    "setupFilesAfterEnv": [
+      "<rootDir>client/setupTests.js"
+    ]
   },
   "dependencies": {
     "@material-ui/core": "^4.2.0",
