+++ server/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const path = require('path');
+const parser = require('body-parser');
 
 const app = express();
 const PORT = 3000;
@@ -11,12 +12,13 @@ const MenuItem = require('./models/MenuItem.js');
 
 
 app.use(express.static(path.join(__dirname, '../client/dist')));
+app.use(parser.json());
 
 app.get('/', function(req, res){
   res.send()
 })
 app.get('/api/menu', (req,res) => {
-  MenuItem.getMenu().then((data)=>res.send(data))
+  MenuItem.getMenu().then((data)=>res.status(202).send(data)).catch(err=>console.log('API Error:' + err))
 })
 
 app.listen(PORT, ()=>{
